Title: 사각형 채우기 3, Time: 140ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240523/사각형 채우기 3/rectangle-fill-3.java
@@ -1,7 +1,7 @@
 import java.util.*;
 
 public class Main {
-    public static int[] square = new int[1001];
+    public static long[] square = new long[1001];
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
         Scanner sc =new Scanner(System.in);
